Add more informative DCHECKs for unreproducible bug.
R= pkasting@chromium.org, georgey@chromium.org
BUG=107675
TEST=unit_tests --gtest_filter=ShortcutsProviderTest.ClassifyAllMatchesInString

Review URL: http://codereview.chromium.org/8949064

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@121943 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: eac449911d0db051162af00537a4e1444b4a391f

+++ chrome/browser/autocomplete/shortcuts_provider_shortcut.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2011 The Chromium Authors. All rights reserved.
+// Copyright (c) 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -79,7 +79,7 @@ ACMatchClassifications SpansFromString(const string16& value);
 // Adds match at the end if and only if its style is different from the last
 // match.
 void AddLastMatchIfNeeded(ACMatchClassifications* matches,
-                          int position,
+                          size_t position,
                           int style);
 }  // namespace shortcuts_provider
